Fix mermaid-cli puppeteer invocation
To get this right, convert Dockerfile to follow puppeteer example a little more
closely, and set up a puppeteer-config file to pass --no-sandbox to the
puppeteer invocation. asciidoctor invocations always pass this as an attribute,
which can be disabled on a per-diagram basis, or by passing an
mermaid-puppeteer-config attribute as one of the asciidoctor-args.